+++ example/TeleDart_example.dart
@@ -14,9 +14,9 @@ void main() {
   teledart.startFetching();
 
   // In case you decided to use webhook.
-  teledart.setupWebhook(envVars['HOST_URL'], envVars['BOT_TOKEN'],
-      io.File(envVars['CERT_PATH']), io.File(envVars['KEY_PATH']));
-  teledart.startFetching(webhook: true);
+  // teledart.setupWebhook(envVars['HOST_URL'], envVars['BOT_TOKEN'],
+  //     io.File(envVars['CERT_PATH']), io.File(envVars['KEY_PATH']));
+  // teledart.startFetching(webhook: true);
 
   // You can listen to messages like this
   teledart.onMessage(entityType: 'bot_command', keyword: 'start').listen(