+++ lib/src/telegram/model.g.dart
@@ -194,6 +194,7 @@ Message _$MessageFromJson(Map<String, dynamic> json) {
           : Chat.fromJson(json['forward_from_chat'] as Map<String, dynamic>),
       forward_from_message_id: json['forward_from_message_id'] as int,
       forward_signature: json['forward_signature'] as String,
+      forward_sender_name: json['forward_sender_name'] as String,
       forward_date: json['forward_date'] as int,
       reply_to_message: json['reply_to_message'] == null
           ? null
@@ -304,6 +305,7 @@ Map<String, dynamic> _$MessageToJson(Message instance) {
   writeNotNull('forward_from_chat', instance.forward_from_chat);
   writeNotNull('forward_from_message_id', instance.forward_from_message_id);
   writeNotNull('forward_signature', instance.forward_signature);
+  writeNotNull('forward_sender_name', instance.forward_sender_name);
   writeNotNull('forward_date', instance.forward_date);
   writeNotNull('reply_to_message', instance.reply_to_message);
   writeNotNull('edit_date', instance.edit_date);
@@ -972,6 +974,7 @@ ChatMember _$ChatMemberFromJson(Map<String, dynamic> json) {
       can_restrict_members: json['can_restrict_members'] as bool,
       can_pin_messages: json['can_pin_messages'] as bool,
       can_promote_members: json['can_promote_members'] as bool,
+      is_member: json['is_member'] as bool,
       can_send_messages: json['can_send_messages'] as bool,
       can_send_media_messages: json['can_send_media_messages'] as bool,
       can_send_other_messages: json['can_send_other_messages'] as bool,
@@ -999,6 +1002,7 @@ Map<String, dynamic> _$ChatMemberToJson(ChatMember instance) {
   writeNotNull('can_restrict_members', instance.can_restrict_members);
   writeNotNull('can_pin_messages', instance.can_pin_messages);
   writeNotNull('can_promote_members', instance.can_promote_members);
+  writeNotNull('is_member', instance.is_member);
   writeNotNull('can_send_messages', instance.can_send_messages);
   writeNotNull('can_send_media_messages', instance.can_send_media_messages);
   writeNotNull('can_send_other_messages', instance.can_send_other_messages);